Create a new testable function `patternToTree`
