+++ src/index.ts
@@ -1,24 +1,4 @@
-import { Lexer } from 'chevrotain';
-
 import expandTree from './expand-tree';
-import Parser from './parser';
-import tokens from './tokens';
+import patternToTree from './pattern-to-tree';
 
-const lexer = new Lexer(tokens);
-const parser = new Parser();
-
-export = (pattern: string) => {
-  const lexingResult = lexer.tokenize(pattern);
-
-  parser.input = lexingResult.tokens;
-
-  const result = parser.tree();
-
-  if (parser.errors.length > 0) {
-    throw new Error(
-      `Parsing error detected!\n${parser.errors[0].message}`,
-    );
-  }
-
-  return expandTree(result);
-};
+export = (pattern: string) => expandTree(patternToTree(pattern));
